Best effort to block particular kind of mutable expressions in Python Cloud Debugger.
The problem is with expressions like x.__setattr__('a', 1). For some reason Python just decides to skip the trace call, which makes calls to primitives like __setattr__ invisible for ImmutabilityTraces. This CL implements a best effort solution (which is by no mean complete). We scan names of a code object and block it if there is a string like "__setattr__". It is possible by bypass it, but it should cover vast majority of the cases when the user runs such expressions by mistake.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=112289038