Modified python agent to include the request log id into the breakpoint labels for GAE Classic applications only.
This is a re-submit of cl/125829312 that was rolled back.
It should work now that we checked in the modified serializer to/from json/protobuf[2|3].

I will send another CL for the updated aeta python test, i prefer having the test get in into a separate CL (so that @emrekultursay can run the plain test the day after we check in against nightly and also the updated one).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=130292452