Refactor auth in the Python agent.
It now uses the google.auth library instead of the deprecated oauth2client.
The project id is retrieved from the credentials, and the project number defaults to the project id if not specified. No flags are needed if application default credentials are available, but the --service_account_json_file, --project_id, and --project_number flags can still override the defaults.
Support for P12 credentials have been removed. They are not mentioned anywhere
in the docs anymore, and oauth2client itself removed the function we used for
P12 credentials in 2016.
https://github.com/google/oauth2client/releases/tag/v2.0.0

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=192151169