feat: add Firebase RTDB backend (#48)
This is a flag-controlled alternative backend for the debug agent to use.  This is provided as a preview of the functionality and will not be released until additional error handling and customization have been added.

This version of the debugger is intended to be used with https://github.com/GoogleCloudPlatform/snapshot-debugger.